extending the macro to work in catkin
The generated header needs to end up in devel/include/moduleName
to match the expectations of the kn build organization and
the catkin devel-space concept. Unfortunately catkin knows
nothing about moduleName.

We had some discussion on a potential reorg in kn, but that would
quickly spiral out of control in terms of effort on NASA sources.
So we decided to just make it to work (tm).